core/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-allauth==65.14.3
django-cors-headers==4.9.0
python-decouple==3.8
whitenoise==6.11.0
bcrypt==5.0.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
asgiref==3.9.1
sqlparse==0.5.3
PyJWT==2.10.1
requests==2.31.0
dj-database-url==2.2.0
django-extensions==3.2.3
